파일 확대보기, 삭제 기능까지 구현 완 / 안드로이드 뷰 갱신 만들어야 함

+++ server_intravel/src/main/java/bitc/fullstack405/server_intravel/service/PhotoService.java
@@ -60,7 +60,17 @@ public class PhotoService {
     return updatePhoto;
   }
 
+  @Transactional
   public void deletePhoto(Long photoId) {
+    PhotoEntity photo = photoRepository.findById(photoId).get();
+
+    String fullPath = photo.getFilePath() + photo.getFileName();
+
+    boolean isFileDeleted = FileUtil.deleteFile(fullPath);
+    if (!isFileDeleted) {
+      throw new IllegalStateException("Failed to delete file: " + fullPath);
+    }
+
     photoRepository.deleteById(photoId);
   }
 }
